refactored DTO's
